VideoCommon: move custom texture data to assets
