+++ Source/Core/VideoCommon/Assets/CustomTextureData.cpp
@@ -1,7 +1,7 @@
 // Copyright 2023 Dolphin Emulator Project
 // SPDX-License-Identifier: GPL-2.0-or-later
 
-#include "VideoCommon/GraphicsModSystem/Runtime/CustomTextureData.h"
+#include "VideoCommon/Assets/CustomTextureData.h"
 
 #include <algorithm>
 #include <cstddef>
