+++ Source/Core/VideoCommon/Assets/CustomTextureData.h
