+++ Source/Core/VideoCommon/CMakeLists.txt
@@ -12,6 +12,8 @@ add_library(videocommon
   Assets/CustomAsset.h
   Assets/CustomAssetLibrary.cpp
   Assets/CustomAssetLibrary.h
+  Assets/CustomTextureData.cpp
+  Assets/CustomTextureData.h
   AsyncRequests.cpp
   AsyncRequests.h
   AsyncShaderCompiler.cpp
@@ -69,8 +71,6 @@ add_library(videocommon
   GraphicsModSystem/Runtime/Actions/ScaleAction.h
   GraphicsModSystem/Runtime/Actions/SkipAction.cpp
   GraphicsModSystem/Runtime/Actions/SkipAction.h
-  GraphicsModSystem/Runtime/CustomTextureData.cpp
-  GraphicsModSystem/Runtime/CustomTextureData.h
   GraphicsModSystem/Runtime/FBInfo.cpp
   GraphicsModSystem/Runtime/FBInfo.h
   GraphicsModSystem/Runtime/GraphicsModAction.h
